feat: implements exceptions to methods

+++ src/main/java/dev/danilo/gymanager/service/UserService.java
@@ -6,7 +6,10 @@ import dev.danilo.gymanager.entity.User;
 import dev.danilo.gymanager.mapper.UserMapper;
 import dev.danilo.gymanager.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,14 +29,23 @@ public class UserService {
 
     public UserResponseDTO save(UserRequestDTO dto) {
         log.info("Saving an User");
+
+        if (repository.existsByEmail(dto.email())) {
+            throw new DataIntegrityViolationException("Email already registered: " + dto.email());
+        }
         return mapper.toDto(repository.save(mapper.toEntity(dto)));
     }
     
     public void delete(String email) {
-        log.info("Stating the process of deleting an user by email: " + email);
+        log.info("Stating the process of deleting an user by email: {}", email);
         UserDetails byEmail = repository.findByEmail(email);
 
-        log.info("User deleted: " + (User) byEmail);
+        if(byEmail == null) {
+            log.error("User not found with email: {}", email);
+            throw new UsernameNotFoundException("User not found with email: " + email);
+        }
+
+        log.info("User deleted: {}", (User) byEmail);
 
         repository.delete((User) byEmail);
     }
